adding new objects objects and code refactoring. update db model -> adding created_at, updated_at

+++ src/main/java/lv/javaguru/java2ToDoApp/businesslogic/api/GetTaskByIdService.java
@@ -5,5 +5,5 @@ import lv.javaguru.java2ToDoApp.domain.Task;
 import java.util.Optional;
 
 public interface GetTaskByIdService {
-    public Optional<Task> getTaskById(Integer id);
+    public Optional<Task> getTaskById(Long id);
 }
