+++ src/main/java/lv/javaguru/java2ToDoApp/businesslogic/api/UpdateTaskService.java
@@ -3,5 +3,5 @@ package lv.javaguru.java2ToDoApp.businesslogic.api;
 import lv.javaguru.java2ToDoApp.businesslogic.impl.Response;
 
 public interface UpdateTaskService {
-    public Response updateTask(Integer id, String title, String done, String dueDate, String priority);
+    public Response updateTask(Long id, String title, String done, String dueDate, String priority);
 }
